sam2/requirements.txt
fastapi
uvicorn
google-cloud-storage
python-dotenv
pillow
torch>=2.3.1
torchvision
numpy
opencv-python-headless
supervision
python-multipart
hydra-core==1.3.2
omegaconf==2.3.0